// src/app/models/Reception.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface Reception{
    CarChickenDeath: number;
    CarNumber: string;
    DriverName: string;
    Execution: number;
    ExitID: string;
    FarmCode: string;
    FinishStaughter: string;
    HourArrival: string;
    HourLeft: string;
    NumberofChicken: number;
    ReceptionDate: string;
    StartStaughter: string;

}